requirements.txt
pyserial
keyboard
grpcio
grpcio-tools
protobuf
hid
crsf-parser